+++ src/components/MarrdownComponents.tsx
@@ -1,0 +1,34 @@
+import Image from "next/image";
+import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
+import { oneDark } from "react-syntax-highlighter/dist/esm/styles/prism";
+
+const components = {
+  code({
+    inline,
+    className,
+    children,
+    ...props
+  }: {
+    inline?: boolean;
+    className?: string;
+    children?: React.ReactNode;
+  }) {
+    const match = /language-(\w+)/.exec(className || "");
+    return !inline && match ? (
+      <SyntaxHighlighter style={oneDark} language={match[1]} PreTag="div" {...props}>
+        {String(children).replace(/\n$/, "")}
+      </SyntaxHighlighter>
+    ) : (
+      <code className={className} {...props}>
+        {children}
+      </code>
+    );
+  },
+  img: ({ src = "", alt, ...props }: { src?: string; alt?: string }) => (
+    <span className="relative block aspect-[3/2] h-auto w-full">
+      <Image src={src ?? ""} alt={alt ?? ""} fill sizes="50vw" className="object-contain" {...props} />
+    </span>
+  ),
+};
+
+export default components;
