+++ src/styles/globals.css
@@ -36,7 +36,7 @@ input:-webkit-autofill {
 }
 
 .prose img {
-  margin: 2rem auto 0;
+  margin: 0 auto;
 }
 
 :root {
